+++ isis/src/tgo/apps/tgocassismos/tgocassismos.h
@@ -1,0 +1,11 @@
+#ifndef tgocassismos_h
+#define tgocassismos_h
+
+#include "Cube.h"
+#include "UserInterface.h"
+
+namespace Isis{
+  extern void tgocassismos(UserInterface &ui);
+}
+
+#endif
